Issue#33 - moved all env to .cfg files and made some final refactoring.
