+++ config/localhost.cfg
@@ -1,2 +1,3 @@
+export PILLAR_ADDRESS=:8080
 #export PILLAR_HOME=
 #export MONGODB_URL=mongodb://user:password@host:port/database?options
